# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Domakin API",
        "description": "API documentation for Domakin platform",
        "contact": {
            "email": "support@domakin.nl"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "https://orange.domakin.nl",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/user/edit-details": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update user profile",
                "operationId": "370355ee8ca1c277b09b5b0a0dcb413e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+31 6 12345678"
                                    },
                                    "password": {
                                        "description": "New password (optional)",
                                        "type": "string"
                                    },
                                    "profileImage": {
                                        "description": "Profile image (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/career/apply": {
            "post": {
                "tags": [
                    "Career"
                ],
                "summary": "Submit a career application",
                "operationId": "f64e3b788b3369f147d2c76d8dea9844",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "position",
                                    "location"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+31 6 12345678"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "viewing_agent"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Amsterdam, Netherlands"
                                    },
                                    "experience": {
                                        "description": "Optional",
                                        "type": "string"
                                    },
                                    "message": {
                                        "description": "Optional",
                                        "type": "string"
                                    },
                                    "resume": {
                                        "description": "Optional resume file (PDF/DOC/DOCX)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Application submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/feedback/list": {
            "get": {
                "tags": [
                    "Feedback"
                ],
                "summary": "List approved feedbacks",
                "operationId": "86b069d772ead07afeede97ea7aa5744",
                "parameters": [
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Language code (optional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/feedback/create": {
            "post": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Create a feedback",
                "operationId": "b1e70977611c867e45febc475c68c803",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "language"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Great service!"
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "en"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Feedback created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}